+++ pkg/ingester/metrics.go
@@ -324,7 +324,6 @@ type tsdbMetrics struct {
 	tsdbBlocksBytes        *prometheus.Desc
 
 	tsdbOOOAppendedSamples *prometheus.Desc
-	tsdbTooOldSamples      *prometheus.Desc
 
 	checkpointDeleteFail    *prometheus.Desc
 	checkpointDeleteTotal   *prometheus.Desc
@@ -512,10 +511,6 @@ func newTSDBMetrics(r prometheus.Registerer) *tsdbMetrics {
 			"cortex_ingester_tsdb_out_of_order_samples_appended_total",
 			"Total number of out of order samples appended.",
 			nil, nil),
-		tsdbTooOldSamples: prometheus.NewDesc(
-			"cortex_ingester_tsdb_too_old_samples_total",
-			"Total number of too old samples.",
-			nil, nil),
 
 		memSeriesCreatedTotal: prometheus.NewDesc(
 			"cortex_ingester_memory_series_created_total",
@@ -578,7 +573,6 @@ func (sm *tsdbMetrics) Describe(out chan<- *prometheus.Desc) {
 	out <- sm.tsdbExemplarsOutOfOrder
 
 	out <- sm.tsdbOOOAppendedSamples
-	out <- sm.tsdbTooOldSamples
 
 	out <- sm.memSeriesCreatedTotal
 	out <- sm.memSeriesRemovedTotal
@@ -631,7 +625,6 @@ func (sm *tsdbMetrics) Collect(out chan<- prometheus.Metric) {
 	data.SendSumOfCounters(out, sm.tsdbExemplarsOutOfOrder, "prometheus_tsdb_exemplar_out_of_order_exemplars_total")
 
 	data.SendSumOfCounters(out, sm.tsdbOOOAppendedSamples, "prometheus_tsdb_head_out_of_order_samples_appended_total")
-	data.SendSumOfCounters(out, sm.tsdbTooOldSamples, "prometheus_tsdb_too_old_samples_total")
 
 	data.SendSumOfCountersPerUser(out, sm.memSeriesCreatedTotal, "prometheus_tsdb_head_series_created_total")
 	data.SendSumOfCountersPerUser(out, sm.memSeriesRemovedTotal, "prometheus_tsdb_head_series_removed_total")